+++ lib/presentation/game/ui/screens/admin_panel_screen.dart
@@ -1,13 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:trivia_app/data/socket_client.dart';
-import 'package:trivia_app/presentation/connected_users/ui/connected_users_ui.dart';
-import 'package:trivia_app/presentation/game_cubit/game_cubit.dart';
-import 'package:trivia_app/presentation/question_panel/ui/finish_button.dart';
+import 'package:trivia_app/presentation/game/cubit/game_cubit.dart';
+import 'package:trivia_app/presentation/game/ui/widgets/connected_users_ui.dart';
 
-import '../../../di/locator.dart';
-import '../../round/round_cubit.dart';
-import '../question_form/question_form.dart';
+import 'package:trivia_app/presentation/game/ui/widgets/finish_button.dart';
+import 'package:trivia_app/presentation/game/ui/widgets/question_view.dart';
+
+import '../../../../di/locator.dart';
+import '../../../question_form/question_form.dart';
 
 class AdminPanelScreen extends StatelessWidget {
   final String gameId;
@@ -73,57 +74,6 @@ class _Body extends StatelessWidget {
           const Expanded(child: ConnectedUsersUi()),
         ],
       ),
-    );
-  }
-}
-
-class QuestionView extends StatelessWidget {
-  const QuestionView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<RoundCubit, RoundState>(
-      builder: (context, state) {
-        final question = state.questionPayload;
-        final connectedUsers = context.read<GameCubit>().state.connectedUsers;
-        return Column(
-          children: [
-            Text(
-              question.question,
-              style: Theme.of(context).textTheme.headline3,
-            ),
-            const SizedBox(height: 16),
-            ...question.answerOptions.map((answer) {
-              final answerCount = state.answers
-                  .where((ans) => ans.userAnswer == answer)
-                  .map((e) => e.getUserIn(connectedUsers))
-                  .toList()
-                  .length;
-              return Row(
-                children: [
-                  Text(answer),
-                  const SizedBox(width: 6),
-                  Text(answerCount.toString())
-                ],
-              );
-            }).toList(),
-            const CountDownUi()
-          ],
-        );
-      },
-    );
-  }
-}
-
-class CountDownUi extends StatelessWidget {
-  const CountDownUi({Key? key}) : super(key: key);
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<RoundCubit, RoundState>(
-      builder: (context, state) {
-        return Text(
-            'Remaining: ${(state.remainingMilliseconds / 1000).toStringAsFixed(2)}');
-      },
     );
   }
 }
