--- lib/presentation/question_panel/cubit/question_panel_cubit.dart
@@ -1,9 +1,0 @@
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'question_panel_state.dart';
-part 'question_panel_cubit.freezed.dart';
-
-class QuestionPanelCubit extends Cubit<QuestionPanelState> {
-  QuestionPanelCubit() : super(const QuestionPanelState.initial());
-}
