--- lib/presentation/question_panel/cubit/question_panel_state.dart
@@ -1,6 +1,0 @@
-part of 'question_panel_cubit.dart';
-
-@freezed
-class QuestionPanelState with _$QuestionPanelState {
-  const factory QuestionPanelState.initial() = _Initial;
-}
